Adding no_rocm tag to \"large\" tests that are failing on the ROCm platform.
Note that these are not regressions per-se. These tests were not being run as part of ROCm CI tests, as only `small` and `medium` tests were being run by default. Now that we are in the process of adding \`large` tests to the ROCM CI run, we have un-covered these failures.

The failures will be analysed and fixed if necessary in a separate PR.